# backend/api_docs.yaml
swagger: '2.0'
info:
  version: "1"
  title: "MusiCloud API"
  description: >-
    This is the REST API documentation for the MusiCloud music mixing platform.
host: "dcumusicloud.com"
basePath: "/api/v1"
...
schemes:
  - "http"
paths:
...
  /users:
    get:
      tags:
      - "users"
      summary: "Get a user's publicly available information."
      description: "By passing the username of the user who's information you want to see, along with your access_token, you will receive back a JSON object with a user's basic profile information."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Header"
          description: "This is where you should set your access_token."
          required: true
          type: "string"
        - in: "query"
          name: "username"
          type: "string"
          required: false
          description: "The username of the user who's info you wish to view."
      responses:
        200:
          description: "User's Info"
          schema:
            type: "object"
            properties:
              followers:
                type: "integer"
                example: 200
              following:
                type: "integer"
                example: 500
              likes:
                type: "integer"
                example: 123
              posts:
                type: "integer"
                example: 100
              profile_pic_url:
                type: "string"
                example: "http://image.fake"
              songs:
                type: "integer"
                example: 321
              username:
                type: "string"
                example: "TestUser"
        401:
          description: "Access_token expired"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Token expired."
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        500:
          description: "Server error"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Server failed to decode token."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
    post:
      tags:
      - "users"
      summary: "Register a new user"
      description: "Creates a new user entry in the Users table. This is step 1 in the authentication process."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New user's info."
        required: true
        schema:
          type: "object"
          properties:
            username:
              type: "string"
              example: "newUser"
            email:
              type: "string"
              example: "newUser@gmail.com"
            password:
              type: "string"
              example: "securePassword1234!"
      responses:
        200:
          description: "User created"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "User created!"
        400:
          description: "Invalid email address format."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Invalid email address."
        409:
          description: "User already exists."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "User already exists!"
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        500:
          description: "Server error"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Error while hashing password."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
    patch:
      tags:
      - "users"
      summary: "Change a user's info"
      description: "Changes the email and or password for user who's access_token has been passed in. This endpoint requires the user to sent there current_password."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "Authorization Header"
        description: "This is where you should set your access_token."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "New info"
        schema:
          type: "object"
          properties:
            password:
              type: "string"
              example: "securePassword1234!"
            email:
              type: "string"
              example: "fake@email.com"
            current_password:
              type: "string"
              example: "1234"
      responses:
        200:
          description: "Details updated"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Email reset, and verification mail sent. Password reset."
        401:
          description: "Bad credentials."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Incorrect password!"
        409:
          description: "Email address already in use"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Duplicate entry"
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        500:
          description: "Server error"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Error while hashing password."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
  /users/reverify:
    post:
      tags:
      - "users"
      summary: "Resend verification email"
      description: "Checks if a given user is in the process of being verified and resends the verification email if they are."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Email of the user looking for a new verification email"
        required: true
        schema:
          type: "object"
          properties:
            email:
              type: "string"
              example: "newUser@gmail.com"
      responses:
        200:
          description: "Resent verification email."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Verification email sent."
        400:
          description: "Some info is bad or is missing!"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Bad request."
        403:
          description: "Request forbidden as the user is already verified."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Already verified."
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
  /users/post:
    post:
      tags:
      - "users"
      summary: "Post a new status update."
      description: "Takes the text and access_token of a given user and makes a post on there account."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "Authorization Header"
        description: "This is where you should set your access_token."
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The message you wish to post."
        required: true
        schema:
          type: "object"
          properties:
            message:
              type: "string"
              example: "Hi my amazing fans!"
      responses:
        200:
          description: "Post made successfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Message posted."
        401:
          description: "User's token has expired"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Token expired."
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        500:
          description: "Failed to decode token."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Server failed to decode token."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
  /users/posts:
    get:
      tags:
      - "users"
      summary: "Get all of the posts made by a particular user."
      description: "This endpoint takes several arguments and returns a subset of posts, ordered chronologically from newest to oldest, made by a particular user. The arguments that can be provided are as follows:\n\n \"username\" = If a 'next_page' or 'back_page' argument is not sent, then this argument is mandatory. This should be set to the username of the user who's posts you are trying to view.\n\n \"posts_per_page\" = If a 'next_page' or 'back_page' argument is not sent, then this an optional argument you may send. This is used to adjust the number of posts you will get back per page of results.\n\n \"current_page\" = If a 'next_page' or 'back_page' argument is not sent, then this is an optional argument you may send. This is used to navigate through each of the pages of results. If this argument is not sent, the default page of results returned will be page 1.\n\n \"next_page\" = If this value is sent, then a back_page value should not be sent. This value should be set to a 'next_page' token. These tokens are returned by standard requests to support infinited scrolling. Making a request with this token requires no other value to be passed as it encodes the user's access token, the page value we want and the number of posts we want per page.\n\n \"back_page\" = If this value is sent, then a next_page value should not be sent. This value should be set to a 'back_page' token. These tokens are returned by standard requests to support infinited scrolling. Making a request with this token requires no other value to be passed as it encodes the user's access token, the page value we want and the number of posts we want per page."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Header"
          description: "This is where you should set your access_token."
          required: true
          type: "string"
        - in: "query"
          name: "username"
          type: "string"
          required: false
          description: "The username of the user who's posts you wish to view."
        - in: "query"
          name: "current_page"
          type: "string"
          required: false
          description: "The page of results you wish to view."
        - in: "query"
          name: "posts_per_page"
          type: "string"
          required: false
          description: "The number of posts you wish to have per page of results."
        - in: "query"
          name: "next_page"
          type: "string"
          required: false
          description: "A next page token to get the next page of results."
        - in: "query"
          name: "back_page"
          type: "string"
          required: false
          description: "A back_page token to get the previous page of results."
      responses:
        200:
          description: "Post made successfully"
          schema:
            type: "object"
            properties:
              back_page:
                type: "string"
                example: "will be null or a token"
              current_page:
                type: "integer"
                example: 1
              next_page:
                type: "string"
                example: "will be null or a token"
              posts:
                type: "array"
                items:
                  type: "array"
                  items:
                    type: "string"
                  example:
                    - "test message"
                    - "Fri, 01 Nov 2019 19:12:27 GMT"
              posts_per_page:
                type: "integer"
                example: 50
              total_pages:
                type: "integer"
                example: 2
        401:
          description: "User's token has expired"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Token expired."
        422:
          description: "Malformed request body."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "A req validation traceback will be here."
        500:
          description: "Failed to decode token."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Server failed to decode token."
        503:
          description: "Server is unavailable"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "MySQL unavailable."
...